fix bug causing gh convenience aliases to incorrectly resolve to releases with hitherto unpopulated assets
